Build lanes on Unix only so macOS and Windows compile again
borg-lanes became an unconditional dependency of the CLI and agent runtime,
and platform CI has failed since: on macOS statvfs fields are narrower than
u64, and on Windows the crate uses flock, fcntl, signals, process groups and
Unix sockets throughout.

Widen the statvfs fields for macOS. Compile borg-lanes only on Unix, and
gate `borg lane`, `borg worktree` and the lane_job, lane_service and
lane_workspace tools behind cfg(unix). Lanes already require Unix process
control, and v0.10.0 shipped Windows builds without them.

+++ crates/borg-lanes/src/workspace/hygiene.rs
@@ -162,7 +162,9 @@ pub fn disk_available(path: &Path) -> Result<u64> {
     );
     // SAFETY: statvfs succeeded and initialized the allocation.
     let stat = unsafe { stat.assume_init() };
-    Ok(stat.f_bavail.saturating_mul(stat.f_frsize))
+    // fsblkcnt_t and c_ulong are narrower than u64 on some Unixes (macOS).
+    #[allow(clippy::unnecessary_cast)]
+    Ok((stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64))
 }
 
 #[cfg(not(unix))]
